pyproject.toml
[tool.poetry]
name = "coreml-bert"
version = "0.1.0"
description = ""
authors = ["Anentropic <ego@anentropic.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
ane-transformers = "^0.1.3"
loguru = "^0.6.0"
huggingface-hub = "^0.13.3"
datasets = "^2.11.0"

[tool.poetry.group.dev.dependencies]
ipython = "^8.12.0"
pdbpp = "^0.10.3"
asitop = "^0.0.23"
huggingface-hub = {extras = ["cli"], version = "^0.13.3"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
